Added Simple HTML FORMS
Now values can be added/deleted using the HTML form interface with Flask.

The dictionary held in todo.json now reorders keys as to keep it sorted in a type of order.

In crud.py the delay function has now been removed as the the UI will be web-based now, not command-line based.